requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.7
idna==3.10
iniconfig==2.0.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.45.3
tomli==2.2.1
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
types-setuptools==75.8.0.20250210
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
